Returning to DHT library after trying the PietteTech DHT library to fix deep sleep issue.
From the ESP8266.com forum I am trying to set the DHTPIN as OUTPUT and set it value to LOW to see if this can fix the problem.

I also added a little more error checking in the data so only valid values are posted to the MQTT broker. Otherwise, a MQTT post is send to the log channel giving the sensorID and the error code.